checkpoint / restore z80s

+++ src/machine.cpp
@@ -144,14 +144,14 @@ size_t Checkpoint::write(bool b) {
 }
 
 size_t Checkpoint::read(uint8_t *buf, size_t len) {
-	for (int i = 0; i < len; i++)
+	for (size_t i = 0; i < len; i++)
 		if (read(buf[i]) == 0)
 			return i;
 	return len;
 }
 
 size_t Checkpoint::write(const uint8_t *buf, size_t len) {
-	for (int i = 0; i < len; i++)
+	for (size_t i = 0; i < len; i++)
 		write(buf[i]);
 	return len;
 }
